requirements.txt
accessible-pygments==0.0.4
alabaster==0.7.13
Babel==2.12.1
click==8.1.3
docutils==0.18.1
ghp-import==2.1.0
greenlet==2.0.2
guzzle-sphinx-theme==0.7.11
imagesize==1.4.1
importlib-metadata==6.7.0
Jinja2==3.1.2
jupyter-book==0.15.1
jupyter-cache==0.6.1
jupyter-contrib-core==0.4.2
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyterlab_myst==2.0.1
jupytext==1.14.7
latexcodec==2.0.1
linkify-it-py==2.0.2
lxml==4.9.3
markdown-it-py==2.2.0
mdit-py-plugins==0.3.5
mdurl==0.1.2
myst-nb==0.17.2
myst-parser==0.18.1
nbclient==0.7.4
pybtex==0.24.0
pybtex-docutils==1.0.2
pydata-sphinx-theme==0.13.3
PyYAML==6.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinx-book-theme==1.0.1
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx-jupyterbook-latex==0.5.2
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.2.1
sphinx-togglebutton==0.3.2
sphinx_design==0.3.0
sphinx_external_toc==0.3.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.17
tabulate==0.9.0
toml==0.10.2
tqdm==4.65.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
uc-micro-py==1.0.2
zipp==3.15.0
